pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "django (>=4.2,<5.0)",
    "djangorestframework (>=3.14.0)",
    "psycopg2 (>=2.9.9)",
    "numpy (==1.26.4)",
    "pandas (>=1.5.3)",
    "scikit-learn (>=1.2.2)",
    "opencv-python (>=4.8.1.78)",
    "pillow (>=10.0.0)",
    "tensorflow-hub (>=0.15.0)",
    "fuzzywuzzy (>=0.18.0)",
    "python-levenshtein (>=0.12.2)",
    "django-haystack (>=3.2.2)",
    "requests (>=2.31.0)",
    "django-allauth (>=0.60.0)",
    "djangorestframework-simplejwt (>=5.3.0)",
    "gunicorn (>=21.2.0)",
    "whitenoise (>=6.6.0)",
    "python-dotenv (>=1.0.0)",
    "transformers (>=4.53.3,<5.0.0)",
    "huggingface (>=0.0.1,<0.0.2)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "paywix (>=2.0.2,<3.0.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)",
    "whoosh (>=2.7.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
